Function transaction_builder::encode_mint_lbr_script

pub fn encode_mint_lbr_script(amount_lbr: u64) -> Script

Mints LBR from the sending account's constituent coins by depositing in the on-chain LBR reserve. Deposits the newly-minted LBR into the sending account. Can be sent by any account that can hold balances for the constituent currencies for LBR and LBR.

Technical Description

Mints amount_lbr LBR from the sending account's constituent coins and deposits the resulting LBR into the sending account.

Events

Successful execution of this script emits three events:

A LibraAccount::SentPaymentEvent with the Coin1 currency code, and a LibraAccount::SentPaymentEvent with the Coin2 currency code on account's LibraAccount::LibraAccount sent_events handle with the amounts for each event being the components amounts of amount_lbr LBR; and

A LibraAccount::ReceivedPaymentEvent on account's LibraAccount::LibraAccount received_events handle with the LBR currency code and amount field equal to amount_lbr.

Parameters

Name	Type	Description

account	&signer	The signer reference of the sending account.

amount_lbr	u64	The amount of LBR (in microlibra) to be created.

Common Abort Conditions

Error Category	Error Reason	Description

Errors::NOT_PUBLISHED	LibraAccount::EPAYER_DOESNT_HOLD_CURRENCY	account doesn't hold a balance in one of the backing currencies of LBR.

Errors::INVALID_ARGUMENT	LBR::EZERO_LBR_MINT_NOT_ALLOWED	amount_lbr passed in was zero.

Errors::LIMIT_EXCEEDED	LBR::ECOIN1	The amount of Coin1 needed for the specified LBR would exceed LBR::MAX_U64.

Errors::LIMIT_EXCEEDED	LBR::ECOIN2	The amount of Coin2 needed for the specified LBR would exceed LBR::MAX_U64.

Errors::INVALID_STATE	Libra::EMINTING_NOT_ALLOWED	Minting of LBR is not allowed currently.

Errors::INVALID_ARGUMENT	LibraAccount::EPAYEE_CANT_ACCEPT_CURRENCY_TYPE	account doesn't hold a balance in LBR.

Errors::LIMIT_EXCEEDED	LibraAccount::EWITHDRAWAL_EXCEEDS_LIMITS	account has exceeded its daily withdrawal limits for the backing coins of LBR.

Errors::LIMIT_EXCEEDED	LibraAccount::EDEPOSIT_EXCEEDS_LIMITS	account has exceeded its daily deposit limits for LBR.

Related Scripts

Script::unmint_lbr