Crate transaction_builder

Enums

ScriptCall
Structured representation of a call into a known Move script.

Functions

encode_add_currency_to_account_script
Summary

encode_add_recovery_rotation_capability_script
Summary

encode_add_to_script_allow_list_script
Summary

encode_add_validator_and_reconfigure_script
Summary

encode_block_prologue_script

encode_burn_script
Summary

encode_burn_txn_fees_script
Summary

encode_cancel_burn_script
Summary

encode_create_child_vasp_account_script
Summary

encode_create_designated_dealer_script
Summary

encode_create_parent_vasp_account_script
Summary

encode_create_recovery_address_script
Summary

encode_create_validator_account_script
Summary

encode_create_validator_operator_account_script
Summary

encode_freeze_account_script
Summary

encode_peer_to_peer_with_metadata_script
Summary

encode_preburn_script
Summary

encode_publish_shared_ed25519_public_key_script
Summary

encode_register_validator_config_script
Summary

encode_remove_validator_and_reconfigure_script
Summary

encode_rotate_authentication_key_script
Summary

encode_rotate_authentication_key_with_nonce_admin_script
Summary

encode_rotate_authentication_key_with_nonce_script
Summary

encode_rotate_authentication_key_with_recovery_address_script
Summary

encode_rotate_dual_attestation_info_script
Summary

encode_rotate_shared_ed25519_public_key_script
Summary

encode_set_validator_config_and_reconfigure_script
Summary

encode_set_validator_operator_script
Summary

encode_set_validator_operator_with_nonce_admin_script
Summary

encode_stdlib_upgrade_transaction
Update WriteSet

encode_tiered_mint_script
Summary

encode_unfreeze_account_script
Summary

encode_update_diem_version_script
Summary

encode_update_dual_attestation_limit_script
Summary

encode_update_exchange_rate_script
Summary

encode_update_minting_ability_script
Summary

get_transaction_name
Returns a user friendly mnemonic for the transaction type if the transaction is for a known, white listed, transaction.